Halifax Curation Script
This is the version of the curation script we are using for https://spineimage.ca/NSHA/site_012/.

* Makes `${SOURCEDATA}` a command line argument

There are some dataset-specific quirks:

* There is a third ID: the internal ID which are kept secret because they are linked to PII, Praxis's IDs (which), and the BIDS IDs; this is a complication that to me seems unnecessary -- we could skip Praxis's IDs, or use them directly as the BIDS IDs, but so it goes. So we need to map one to the other, and I added a parameter for that. I did this with `grep`, which isn't that reliable but was enough for a first draft.
*  instead of `${sourcedata}/${ID}`, the source data are in `D:\$ID $NAME\PAT01\DICOM`

Because of the quriks I don't think this should be merged, but I want to keep this branch handy. Maybe we can figure out something. For example, perhaps the mapping from RHSCIR IDs to Praxis(?) IDs can be a supplementary script that generates a temporary participants.tsv not to be committed with the extra column in it.